Moved the drag and drop into scope
