+++ src/components/ParamField.vue
@@ -1,37 +1,45 @@
 <template>
     <div>
-        <p>{{ param.displayName }}</p>
+        <p
+            v-if="!param.hideDescription && param.description"
+            class="param-description">
+            {{ param.description }}
+        </p>
         <div class="input-container">
-            <input
-                v-if="param.type === ParamType.BOOLEAN"
-                type="checkbox"
-                v-bind:id="paramName"
-                v-bind:checked="param.value === 'true'"
-                v-on:input="updateBoolean($event)" />
-            <input
-                v-else-if="param.type === ParamType.COLOR"
-                type="color"
-                v-bind:id="paramName"
-                v-bind:value="`${param.value}`"
-                v-on:input="updateString($event)" />
-            <select
-                v-else-if="param.type === ParamType.ENUM"
-                v-bind:id="paramName"
-                v-bind:value="`${param.value}`"
-                v-on:input="updateString($event)">
-                <option
-                    v-for="(value, name) in getEnumeration(param.from)"
-                    v-bind:key="name"
-                    v-bind:value="value">
-                    {{ name }}
-                </option>
-            </select>
-            <input
-                v-else
-                v-bind:id="paramName"
-                v-bind:class="!status.isValid() ? 'error-field' : ''"
-                v-bind:value="`${param.value}`"
-                v-on:input="updateString($event)" />
+            <label
+                >{{ param.displayName }}
+                <input
+                    v-if="param.type === ParamType.BOOLEAN"
+                    type="checkbox"
+                    v-bind:id="paramName"
+                    v-bind:checked="param.value === 'true'"
+                    v-on:input="updateBoolean($event)" />
+                <input
+                    v-else-if="param.type === ParamType.COLOR"
+                    type="color"
+                    v-bind:id="paramName"
+                    v-bind:value="`${param.value}`"
+                    v-on:input="updateString($event)" />
+                <select
+                    v-else-if="param.type === ParamType.ENUM"
+                    v-bind:id="paramName"
+                    v-bind:value="`${param.value}`"
+                    v-on:input="updateString($event)">
+                    <option
+                        v-for="(value, name) in getEnumeration(param.from)"
+                        v-bind:key="name"
+                        v-bind:value="value">
+                        {{ name }}
+                    </option>
+                </select>
+                <input
+                    v-else
+                    type="text"
+                    v-bind:id="paramName"
+                    v-bind:class="!status.isValid() ? 'error-field' : ''"
+                    v-bind:value="`${param.value}`"
+                    v-on:input="updateString($event)" />
+            </label>
 
             <div
                 class="desc-tooltip"
@@ -45,9 +53,6 @@
             v-for="error in status.errors"
             v-bind:key="error">
             {{ error }}
-        </p>
-        <p v-if="param.hideDescription && param.description">
-            {{ param.description }}
         </p>
     </div>
 </template>
@@ -104,9 +109,53 @@
 </script>
 
 <style scoped lang="scss">
+    label {
+        font-size: 12px;
+        width: 100%;
+    }
+
+    input {
+        &[type='text'] {
+            border: none;
+            border-bottom: 1.5px solid var(--ns-color-black);
+            font-size: 14px;
+            padding: 6px 8px 6px 8px;
+            width: 100%;
+
+            &:focus {
+                border-bottom: 1.5px solid var(--ns-color-primary);
+                outline: none;
+            }
+        }
+    }
+
+    select {
+        border: 1px solid var(--ns-color-black);
+        background-color: var(--ns-color-white);
+        display: block;
+        width: 100%;
+        padding: 8px;
+        margin-top: 4px;
+
+        &:focus {
+            outline: 3px solid var(--ns-color-primary);
+            border: none;
+        }
+    }
+
+    .param-description {
+        font-size: 12px;
+        color: var(--ns-color-grey);
+        margin-bottom: 8px;
+        margin-top: 0px;
+    }
+
     .input-container {
         display: flex;
         flex-direction: row;
+        position: relative;
+
+        margin-bottom: 32px;
     }
 
     .error-field {
@@ -114,24 +163,33 @@
     }
 
     .desc-tooltip {
-        position: relative;
+        position: absolute;
         display: inline-block;
         cursor: default;
+
+        .material-icons-sharp {
+            font-size: 16px;
+            color: var(--ns-color-grey);
+        }
+
+        right: 4px;
+        bottom: 4px;
     }
 
     .desc-tooltip .desc-tooltip-text {
         visibility: hidden;
         width: 240px;
-        background-color: black;
-        color: #fff;
-        text-align: center;
-        border-radius: 6px;
-        padding: 5px 0;
+        background-color: var(--ns-color-white);
+        color: var(--ns-color-black);
+        text-align: left;
+        border: 1px solid var(--ns-color-black);
+        padding: 8px;
+        font-size: 12px;
 
         position: absolute;
         z-index: 1;
         bottom: 125%;
-        left: 50%;
+        right: 0;
         margin-left: -120px;
 
         opacity: 0;
