+++ src/router/index.ts
@@ -1,6 +1,5 @@
 import Gallery from '../views/Gallery.vue'
 import Scope from '../views/Scope.vue'
-import DragAndDrop from '../views/DragAndDrop.vue'
 import {createRouter, createWebHistory} from 'vue-router'
 
 const router = createRouter({
@@ -15,11 +14,6 @@ const router = createRouter({
             path: '/gallery',
             name: 'gallery',
             component: Gallery,
-        },
-        {
-            path: '/dragAndDrop',
-            name: 'dragAndDrop',
-            component: DragAndDrop,
         },
     ],
 })
